requirements.txt
torch
transformers
Pillow
numpy
scikit-learn
faiss-cpu
matplotlib
tqdm
streamlit 